Adjust the topbar layout for the spaces overview
